Refactor shifts
